pyproject.toml
[project]
name = "obsidiantime"
version = "0.1.0"
description = ""
authors = [
    {name = "XetPy1030",email = "xetpy1030@mail.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-crispy-forms (>=2.4,<3.0)",
    "django-filter (>=25.1,<26.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg-binary (>=3.2.9,<4.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "ruff (>=0.12.4,<0.13.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "crispy-bootstrap4 (>=2025.6,<2026.0)",
    "django-tinymce (>=4.1.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
